+++ Jarvis.Framework/ProjectionEngine/Atomic/Support/AtomicReadmodelProjectorHelperFactory.cs
@@ -21,16 +21,7 @@ namespace Jarvis.Framework.Kernel.ProjectionEngine.Atomic.Support
         {
             var genericType = typeof(AtomicReadmodelProjectorHelper<>);
             var closedType = genericType.MakeGenericType(new Type[] { atomicReadmodelType });
-            try
-            {
-
-                return (IAtomicReadmodelProjectorHelper)_kernel.Resolve(closedType);
-            }
-            catch (Exception ex)
-            {
-
-                throw;
-            }
+            return (IAtomicReadmodelProjectorHelper)_kernel.Resolve(closedType);
         }
     }
 }